Split rule-pack.ts into multiple files
